correct Routes.rb modify year and price in cars.rb and car_spec.rb modify test in index_spec.rb

+++ spec/models/car_spec.rb
@@ -4,7 +4,7 @@ RSpec.describe Car, type: :model do
     it { is_expected.to have_db_column(:model).of_type(:string) }
     it { is_expected.to have_db_column(:colour).of_type(:string) }
     it { is_expected.to have_db_column(:year).of_type(:integer) }
-    it { is_expected.to have_db_column(:price).of_type(:float) }
+    it { is_expected.to have_db_column(:price).of_type(:integer) }
   end
 
   describe 'Factory' do
